refactor(web): switch confirmation modal in mission to a generic component
