+++ web/src/pages/EditMission/CreateMission.tsx
@@ -19,32 +19,7 @@ import Feedbacks from '../../component/Feedback';
 import TopBar from '../../component/SideBar/TopBar';
 import SideBar from '../../component/SideBar/SideBar';
 import config from '../../config';
-
-type InputSizes = 'small' | 'medium' | 'large';
-
-type InputProps = {
-    label: string;
-    labelState: any;
-    setLabel: React.Dispatch<React.SetStateAction<string>>;
-    size: InputSizes;
-};
-
-function Input({ label, labelState, setLabel, size }: InputProps) {
-    return (
-        <div className={`input input-${size}`}>
-            <label htmlFor={`input-${label}`} className="input-label">
-                {label}
-            </label>
-            <input
-                id={`input-${label}`}
-                type="text"
-                required
-                value={labelState}
-                onChange={(e) => setLabel(e.target.value)}
-            />
-        </div>
-    );
-}
+import { Input } from './EditMission';
 
 export default function CreateMission() {
     const [Title, setTitle] = useState('');
